+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 ### Added
 - Add `MyAccountsEndpoint.get` and `MyAccount.idpDomain`
+- Add `IdentityProvider` resource and `Account.identityProvider`
 
 ## [1.5.0] - 2021-11-09
 ### Added